Split locate into seperate modules
