+++ ichnaea/locate/location_provider.py
@@ -28,6 +28,9 @@ from ichnaea.models import (
     Wifi,
     WifiLookup,
 )
+from ichnaea.locate.location import PositionLocation, CountryLocation
+from ichnaea.locate.base import StatsLogger
+
 
 # parameters for wifi clustering
 MAX_WIFI_CLUSTER_KM = 0.5
@@ -64,168 +67,6 @@ def estimate_accuracy(lat, lon, points, minimum):
     if accuracy is not None:
         accuracy = float(accuracy)
     return max(accuracy, minimum)
-
-
-def map_data(data, client_addr=None):
-    """
-    Transform a geolocate API dictionary to an equivalent search API
-    dictionary.
-    """
-    mapped = {
-        'geoip': None,
-        'radio': data.get('radioType', None),
-        'cell': [],
-        'wifi': [],
-    }
-    if client_addr:
-        mapped['geoip'] = client_addr
-
-    if not data:
-        return mapped
-
-    if 'cellTowers' in data:
-        for cell in data['cellTowers']:
-            new_cell = {
-                'mcc': cell['mobileCountryCode'],
-                'mnc': cell['mobileNetworkCode'],
-                'lac': cell['locationAreaCode'],
-                'cid': cell['cellId'],
-            }
-            # If a radio field is populated in any one of the cells in
-            # cellTowers, this is a buggy geolocate call from FirefoxOS.
-            # Just pass on the radio field, as long as it's non-empty.
-            if 'radio' in cell and cell['radio'] != '':
-                new_cell['radio'] = cell['radio']
-            mapped['cell'].append(new_cell)
-
-    if 'wifiAccessPoints' in data:
-        mapped['wifi'] = [{
-            'key': wifi['macAddress'],
-        } for wifi in data['wifiAccessPoints']]
-
-    return mapped
-
-
-class AbstractLocation(object):
-    """A location returned by a location provider."""
-
-    def __init__(self, source=None,
-                 lat=None, lon=None, accuracy=None,
-                 country_code=None, country_name=None, query_data=True):
-        self.source = source
-        self.lat = self._round(lat)
-        self.lon = self._round(lon)
-        self.accuracy = self._round(accuracy)
-        self.country_code = country_code
-        self.country_name = country_name
-        self.query_data = query_data
-
-    def _round(self, value):
-        if value is not None:
-            value = round(value, DEGREE_DECIMAL_PLACES)
-        return value
-
-    def found(self):  # pragma: no cover
-        """Does this location include any location data?"""
-        raise NotImplementedError
-
-    def agrees_with(self, other_location):  # pragma: no cover
-        """Does this location match the position of the other location?"""
-        raise NotImplementedError
-
-    def accurate_enough(self):  # pragma: no cover
-        """Is this location accurate enough to return it?"""
-        raise NotImplementedError
-
-    def more_accurate(self, other_location):  # pragma: no cover
-        """Is this location better than the passed in location?"""
-        raise NotImplementedError
-
-
-class PositionLocation(AbstractLocation):
-    """The location returned by a position query."""
-
-    def found(self):
-        return None not in (self.lat, self.lon)
-
-    def agrees_with(self, other_location):
-        dist = distance(
-            other_location.lat, other_location.lon, self.lat, self.lon) * 1000
-        return dist <= other_location.accuracy
-
-    def accurate_enough(self):
-        # For position data we currently always want to continue.
-        return False
-
-    def more_accurate(self, other_location):
-        """
-        Are we more accurate than the passed in other_location and fit into
-        the other_location's range?
-        """
-        if not self.found():
-            return False
-        if not other_location.found():
-            return True
-        if self.source < other_location.source:
-            return True
-        return (
-            self.agrees_with(other_location) and
-            self.accuracy < other_location.accuracy)
-
-
-class CountryLocation(AbstractLocation):
-    """The location returned by a country query."""
-
-    def found(self):
-        return None not in (self.country_code,  self.country_name)
-
-    def agrees_with(self, other_location):  # pragma: no cover
-        return self.country_code == other_location.country_code
-
-    def accurate_enough(self):
-        if self.found():
-            return True
-        return False
-
-    def more_accurate(self, other_location):
-        if not self.found():
-            return False
-        if not other_location.found():
-            return True
-        if self.source < other_location.source:  # pragma: no cover
-            return True
-        return False  # pragma: no cover
-
-
-class StatsLogger(object):
-
-    def __init__(self, api_key_name, api_key_log, api_name):
-        """
-        A StatsLogger sends counted and timed named statistics to
-        a statistic aggregator client.
-
-        :param api_key_name: Human readable API key name
-            (for example 'test_1')
-        :type api_key_name: str
-        :param api_key_log: Gather additional API key specific stats?
-        :type api_key_log: bool
-        :param api_name: Name of the API, used as stats prefix
-            (for example 'geolocate')
-        :type api_name: str
-        """
-        self.api_key_name = api_key_name
-        self.api_key_log = api_key_log
-        self.api_name = api_name
-        self.raven_client = get_raven_client()
-        self.stats_client = get_stats_client()
-
-    def stat_count(self, stat):
-        self.stats_client.incr('{api}.{stat}'.format(
-            api=self.api_name, stat=stat))
-
-    def stat_time(self, stat, count):
-        self.stats_client.timing('{api}.{stat}'.format(
-            api=self.api_name, stat=stat), count)
 
 
 class AbstractLocationProvider(StatsLogger):
@@ -726,126 +567,3 @@ class PositionGeoIPLocationProvider(GeoIPLocationProvider):
 
 class CountryGeoIPLocationProvider(GeoIPLocationProvider):
     location_type = CountryLocation
-
-
-class AbstractLocationSearcher(StatsLogger):
-    """
-    An AbstractLocationSearcher will use a collection of LocationProvider
-    classes to attempt to identify a user's location. It will loop over them
-    in the order they are specified and use the most accurate location.
-    """
-    # First we attempt a "zoom-in" from cell-lac, to cell
-    # to wifi, tightening our estimate each step only so
-    # long as it doesn't contradict the existing best-estimate.
-
-    provider_classes = ()
-
-    def __init__(self, db_sources, *args, **kwargs):
-        super(AbstractLocationSearcher, self).__init__(*args, **kwargs)
-
-        self.all_providers = [
-            cls(
-                db_sources[cls.db_source_field],
-                api_key_log=self.api_key_log,
-                api_key_name=self.api_key_name,
-                api_name=self.api_name,
-            ) for cls in self.provider_classes]
-
-    def search_location(self, data):
-        best_location = None
-        best_location_provider = None
-        all_locations = defaultdict(deque)
-
-        for provider in self.all_providers:
-            provider_location = provider.locate(data)
-            all_locations[provider.log_group].appendleft(
-                (provider, provider_location))
-
-            if (
-                best_location is None or
-                provider_location.more_accurate(best_location)
-            ):
-                # If this location is more accurate than our previous one,
-                # we'll use it.
-                best_location = provider_location
-                best_location_provider = provider
-
-            if best_location.accurate_enough():
-                # Stop the loop, if we have a good quality location.
-                break
-
-        if not best_location.found():
-            self.stat_count('miss')
-        else:
-            best_location_provider.log_hit()
-
-        # Log a hit/miss metric for the first data source for
-        # which the user provided sufficient data
-        for log_group in ('wifi', 'cell', 'geoip'):
-            group_locations = all_locations[log_group]
-            if any([l.query_data for (p, l) in group_locations]):
-                # Claim a success if at least one location for a logging
-                # group was a success.
-                first_provider, first_location = group_locations[0]
-                found_provider = None
-                for (provider, location) in group_locations:
-                    if location.found():
-                        found_provider = provider
-                        break
-                if found_provider:
-                    found_provider.log_success()
-                else:
-                    first_provider.log_failure()
-                break
-
-        return best_location
-
-    def prepare_location(self, country, location):  # pragma: no cover
-        raise NotImplementedError()
-
-    def search(self, data):
-        """Provide a type specific search location or return None."""
-        location = self.search_location(data)
-        if location.found():
-            return self.prepare_location(location)
-        return None
-
-
-class PositionSearcher(AbstractLocationSearcher):
-    """
-    A PositionSearcher will return a position defined by a latitude,
-    a longitude and an accuracy in meters.
-    """
-
-    provider_classes = (
-        PositionGeoIPLocationProvider,
-        OCIDCellAreaLocationProvider,
-        CellAreaLocationProvider,
-        OCIDCellLocationProvider,
-        CellLocationProvider,
-        WifiLocationProvider,
-    )
-
-    def prepare_location(self, location):
-        return {
-            'lat': location.lat,
-            'lon': location.lon,
-            'accuracy': location.accuracy,
-        }
-
-
-class CountrySearcher(AbstractLocationSearcher):
-    """
-    A CountrySearcher will return a country name and code.
-    """
-
-    provider_classes = (
-        CellCountryProvider,
-        CountryGeoIPLocationProvider,
-    )
-
-    def prepare_location(self, location):
-        return {
-            'country_name': location.country_name,
-            'country_code': location.country_code,
-        }
